codealpha_Object_Detection_&_Tracking/requirements.txt
# Object Detection & Tracking Dependencies
# Core
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# Detection (YOLOv8)
ultralytics>=8.0.0

# Tracking
filterpy>=1.4.5
scipy>=1.10.0
lap>=0.4.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
